+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-BJKjN-x_.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-e-cIoext.js";import{_ as a}from"./mermaid.core-wofKSK4y.js";import"./chunk-5VM5RSS4-Crcqyf6T.js";import"./chunk-XXDRQBXY-B84GkPPn.js";import"./chunk-POPQ4Y6H-0DHVCRJc.js";import"./chunk-F27PBJKO-C_Cg1i5a.js";import"./channel-DUAvSgvX.js";import"./index-CFqUK3fC.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};