+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-CSOWyVP8.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-BWcr5MX3.js";import{_ as a}from"./mermaid.core-wrSwZIe3.js";import"./chunk-5VM5RSS4-CwD-F4n8.js";import"./chunk-XXDRQBXY-BildCRzS.js";import"./chunk-POPQ4Y6H-DC3gSFnL.js";import"./chunk-F27PBJKO-jfOnkX1Q.js";import"./channel-DY8RFHwJ.js";import"./index-DnC2mfcm.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};